Replace unused functionality. Minor updates
