+++ FlowCrypt/Controllers/SideMenu/Menu/Models/FolderViewModel.swift
@@ -45,7 +45,7 @@ extension FolderViewModel {
     } 
 }
 
-extension MenuNode.Input {
+extension MenuCellNode.Input {
     init(_ viewModel: FolderViewModel) {
         self.init(attributedText: viewModel.name.attributed(.regular(17)), image: viewModel.image)
     }
