+++ FlowCryptCommon/UIVIewExtensions.swift
@@ -1,0 +1,115 @@
+//
+// © 2017-2019 FlowCrypt Limited. All rights reserved.
+//
+
+import UIKit
+
+public extension UIView {
+    @discardableResult
+    func bordered(color: UIColor, width: CGFloat) -> Self {
+        layer.borderColor = color.cgColor
+        layer.borderWidth = width
+        return self
+    }
+
+    @discardableResult
+    func cornered(_ cornerRadius: CGFloat) -> Self {
+        layer.cornerRadius = cornerRadius
+        return self
+    }
+}
+
+public func borderStyle(color: UIColor, width: CGFloat) -> (UIView) -> Void {
+  return {
+    $0.layer.borderColor = color.cgColor
+    $0.layer.borderWidth = width
+  }
+}
+
+public extension UIView {
+    func constraintsToEdges(to guide: UILayoutGuide) {
+        translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            leadingAnchor.constraint(equalTo: guide.leadingAnchor),
+            trailingAnchor.constraint(equalTo: guide.trailingAnchor),
+            topAnchor.constraint(equalTo: guide.topAnchor),
+            bottomAnchor.constraint(equalTo: guide.bottomAnchor),
+        ])
+    }
+
+    func constrainToEdges(_ subview: UIView, insets: UIEdgeInsets = .zero) {
+        subview.translatesAutoresizingMaskIntoConstraints = false
+
+        let topContraint = NSLayoutConstraint(
+            item: subview,
+            attribute: .top,
+            relatedBy: .equal,
+            toItem: self,
+            attribute: .top,
+            multiplier: 1.0,
+            constant: insets.top
+        )
+
+        let bottomConstraint = NSLayoutConstraint(
+            item: subview,
+            attribute: .bottom,
+            relatedBy: .equal,
+            toItem: self,
+            attribute: .bottom,
+            multiplier: 1.0,
+            constant: -insets.bottom
+        )
+
+        let leadingContraint = NSLayoutConstraint(
+            item: subview,
+            attribute: .leading,
+            relatedBy: .equal,
+            toItem: self,
+            attribute: .leading,
+            multiplier: 1.0,
+            constant: insets.left
+        )
+
+        let trailingContraint = NSLayoutConstraint(
+            item: subview,
+            attribute: .trailing,
+            relatedBy: .equal,
+            toItem: self,
+            attribute: .trailing,
+            multiplier: 1.0,
+            constant: -insets.right
+        )
+
+        addConstraints([
+            topContraint,
+            bottomConstraint,
+            leadingContraint,
+            trailingContraint,
+        ])
+    }
+
+    func constrainToBorders(_ subview: UIView, insets: UIEdgeInsets = .zero) {
+        subview.translatesAutoresizingMaskIntoConstraints = false
+        subview.leftAnchor.constraint(equalTo: leftAnchor, constant: insets.left).isActive = true
+        subview.rightAnchor.constraint(equalTo: rightAnchor, constant: -insets.right).isActive = true
+        subview.topAnchor.constraint(equalTo: topAnchor, constant: insets.top).isActive = true
+        subview.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -insets.bottom).isActive = true
+    }
+}
+
+
+// MARK: - UITextField
+public extension UITextField {
+    func setTextInset(_ left: CGFloat = 7) {
+        leftView = UIView(frame: CGRect(x: 0, y: 0, width: left, height: frame.size.height))
+        leftViewMode = .always
+    }
+
+    func setTextInsets(_ insets: UIEdgeInsets) {
+        leftView = UIView(frame: CGRect(x: 0, y: 0, width: insets.left, height: frame.size.height))
+        leftViewMode = .always
+
+        rightView = UIView(frame: CGRect(x: 0, y: 0, width: insets.right, height: frame.size.height))
+        rightViewMode = .always
+    }
+}
