+++ FlowCryptUI/Resources/Helpers.swift
@@ -13,7 +13,7 @@ public func testAttributedText() -> NSAttributedString {
     return NSAttributedString(
         string: String((5...count).compactMap { _ in "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789".randomElement() }),
         attributes: [
-            NSAttributedString.Key.font : UIFont.boldSystemFont(ofSize: 13),
+            NSAttributedString.Key.font : UIFont.boldSystemFont(ofSize: 14),
             NSAttributedString.Key.foregroundColor : UIColor.black
         ]
     )
